Make integer narrowings explicit in framework code (-Wconversion)
Add explicit casts at the size_t/int/long to fixed-width narrowing sites flagged
by -Wconversion/-Wsign-conversion in hand-written framework code: ASNodeQueries
index/count, ArrayTools array-length to OpcUa_UInt32, FreeVariablesEngine value
narrowing, DSourceVariableThreadPool and CalculatedVariablesEngine sign casts,
and ASNodeManager namespace-index to OpcUa_UInt16. Behavior-preserving; values
are bounded; casts document intent.

+++ AddressSpace/include/ASNodeQueries.h
@@ -196,7 +196,7 @@ namespace AddressSpace
         {
             ABORT_MESSAGE(CONCAT3(" REGEX Expression is wrong:",pattern,e.what()));
         }
-        int numReferenced = findAllByRegex<T> (nm, nm->getObjectsFolderNode(), OpcUa_NodeClass_Object, expression, storage);
+        int numReferenced = static_cast<int>(findAllByRegex<T> (nm, nm->getObjectsFolderNode(), OpcUa_NodeClass_Object, expression, storage));
         int numUnreferenced = 0;
         auto& unreferencedNodes = nm->getUnreferencedNodes();
         for (auto it = unreferencedNodes.begin(); it!=unreferencedNodes.end(); it++)
@@ -222,7 +222,7 @@ namespace AddressSpace
                 }
             }
         }
-        return numReferenced + numUnreferenced;
+        return static_cast<unsigned int>(numReferenced + numUnreferenced);
 
     }
